requirements.txt
yfinance>=0.2.36
pandas>=2.2.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.11.1
requests>=2.31.0
slack-sdk>=3.26.0
pyyaml>=6.0
numpy>=1.24.0
sec-edgar-downloader>=4.3.0
